backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.29
pandas==2.2.1
python-multipart==0.0.9
anthropic>=0.23.1
python-dotenv==1.0.1
aiofiles==23.2.1
openpyxl==3.1.2
